requirements.txt
aniso8601==8.0.0
arrow==0.15.2
bitso-py==3.0.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cloudinary==1.17.0
Django==2.2.5
Flask==1.1.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.7
future==0.17.1
gunicorn==19.9.0
idna==2.8
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mock==3.0.5
Pillow==6.1.0
psycopg2==2.9.1
python-dateutil==2.8.0
pytz==2019.2
qrcode==6.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
sqlparse==0.3.0
urllib3==1.25.5
websocket-client==0.40.0
Werkzeug==0.16.0
